repeated keyword argument Fixes #1
